Add checklist actions to page controller.
The HTML pages will be used mainly to show the react app, so we can
discard the CRUD operation pages that are automatically generated.

We're only interested in creating, showing, editing and deleting
checklists. The dedicated /new, /form endpoints are not needed. Also
most controller actions do not need to retrieve the model information -
that will be done by the frontend.

This also required hacking around with the templates - fixing model data
lookup, fixing router paths, etc.