Different behaviour for Checklist Edit and Execute
Checklists can be edited, where the title and items can be changed, or
executed, where the checbox state can be changed.

Accomplished this different by passing different dispatch function to
the react components. Refactored the API interface code so that the
checklist/ and execute/ JSON APIs can be hit seperately.

The different behaviour is triggered by the url path - if it's of the
form /checklist/:id/execute/execution_id then the behaviour is changed.